+++ libs/ui/src/index.ts
@@ -2,6 +2,7 @@ export * from './lib/containers/Card';
 export * from './lib/containers/FloatingCard';
 
 export * from './lib/display/Text';
+export * from './lib/display/Tooltip';
 
 export * from './lib/forms/Button';
 export * from './lib/forms/Checkbox';
@@ -15,6 +16,7 @@ export * from './lib/navigation/Arrow';
 
 export * from './lib/notification/infoPanel';
 export * from './lib/notification/Toast';
+export * from './lib/notification/Modal';
 
 export { default as themeGuiaBolso } from './designSystem/tokens/guiaBolso';
 export { default as themeConnect } from './designSystem/tokens/connect';